requirements.txt
amqp==5.3.1
asgiref==3.9.1
astroid==3.3.11
attrs==25.3.0
billiard==4.2.2
black==25.1.0
celery==5.5.3
certifi==2025.7.14
charset-normalizer==3.4.4
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
configparser==7.2.0
dill==0.4.0
Django==5.2.4
django-crispy-forms==2.4
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
isort==6.0.1
kombu==5.5.4
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.3.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.52
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
pyee==13.0.0
pylint==3.3.7
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
selenium==4.35.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tomlkit==0.13.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
websocket-client==1.8.0
wheel==0.45.1
wsproto==1.2.0
xlrd==2.0.2
